Moves redux actions and reducers to App-level subfolder
