+++ client/Concepts/Belcurv/src/store/store.js
@@ -1,7 +1,7 @@
 /* ================================= SETUP ================================= */
 
 import { createStore } from 'redux';
-import rootReducer from './reducers';
+import rootReducer from '../reducers';
 
 /* ================================= STORE ================================= */
